requirements.txt
imagehash
pillow
send2trash